requirements.txt
fastapi
uvicorn
python-dotenv
google-api-python-client
websockets
